[#4400] refactor program.jsx into router & overview
