--- akvo/rsr/spa/app/modules/program/program.jsx
@@ -1,144 +1,0 @@
-/* global window, document */
-import React, { useState, useEffect } from 'react'
-import { connect } from 'react-redux'
-import { Collapse, Icon, Spin, Tabs, Select } from 'antd'
-import classNames from 'classnames'
-import { Route, Link, Redirect } from 'react-router-dom'
-import { useTranslation } from 'react-i18next'
-import './styles.scss'
-import Result from './result'
-import Hierarchy from '../hierarchy/hierarchy'
-import Editor from '../editor/editor'
-import api from '../../utils/api'
-import Reports from '../reports/reports'
-import countriesDict from '../../utils/countries-dict'
-import StickyClass from './sticky-class'
-import * as actions from '../editor/actions'
-import Approvals from '../approvals/approvals'
-
-const { Panel } = Collapse
-const { TabPane } = Tabs
-const { Option } = Select
-
-const ExpandIcon = ({ isActive }) => (
-  <div className={classNames('expander', { isActive })}>
-    <Icon type="down" />
-  </div>
-)
-
-const Program = ({ match: {params}, userRdr, ...props }) => {
-  const { t } = useTranslation()
-  const [results, setResults] = useState([])
-  const [title, setTitle] = useState('')
-  const [loading, setLoading] = useState(true)
-  const [countryOpts, setCountryOpts] = useState([])
-  const [countryFilter, setCountryFilter] = useState([])
-  const initiate = () => {
-    setLoading(true)
-    if (params.projectId !== 'new') {
-      api.get(`/project/${params.projectId}/results`)
-        .then(({ data }) => {
-          setResults(data.results.map(it => ({ ...it, indicators: [] })))
-          setTitle(data.title)
-          props.setProjectTitle(data.title)
-          document.title = `${data.title} | Akvo RSR`
-          setLoading(false)
-          // collect country opts
-          const opts = []
-          data.results.forEach(result => {
-            result.countries.forEach(opt => { if (opts.indexOf(opt) === -1) opts.push(opt) })
-          })
-          setCountryOpts(opts)
-        })
-    } else {
-      setLoading(false)
-    }
-  }
-  useEffect(initiate, [params.projectId])
-  const handleResultChange = (index) => {
-    if(index != null){
-      window.scroll({ top: 142 + index * 88, behavior: 'smooth'})
-    }
-  }
-  const handleCountryFilter = (value) => {
-    setCountryFilter(value)
-  }
-  const filterCountry = (filterValue) => (item) => {
-    if(filterValue.length === 0) return true
-    let index = 0
-    let found = false
-    while(filterValue.length > index){
-      if(item.countries.indexOf(filterValue[index]) !== -1) {
-        found = true; break
-      }
-      index += 1
-    }
-    return found
-  }
-  const canEdit = userRdr.programs && userRdr.programs.find(it => it.id === params.projectId && it.canEditProgram) !== -1
-  let _title = props.title
-  if(!_title && title) _title = title
-  else if(!_title) _title = t('Untitled program')
-  return (
-    <div className="program-view">
-      <Route path="/programs/:id/:view?" render={({ match }) => {
-        const view = match.params.view ? match.params.view : ''
-        return [
-          <header className={classNames('main-header', { editor: match.params.view === 'editor' })}>
-            <h1>{!loading && _title}</h1>
-          </header>,
-          <Tabs size="large" activeKey={view}>
-            {(results.length > 0 || !match.params.view) && <TabPane tab={<Link to={`/programs/${params.projectId}`}>Overview</Link>} key="" />}
-            <TabPane tab={<Link to={`/programs/${params.projectId}/approvals`}>Approvals</Link>} key="approvals" />
-            <TabPane disabled={!canEdit} tab={<Link to={`/programs/${params.projectId}/editor`}>Editor</Link>} key="editor" />
-            <TabPane tab={<Link to={`/programs/${params.projectId}/hierarchy`}>Hierarchy</Link>} key="hierarchy" />
-            <TabPane tab={<Link to={`/programs/${params.projectId}/reports`}>Reports</Link>} key="reports" />
-          </Tabs>
-        ]
-      }} />
-      {loading && <div className="loading-container"><Spin indicator={<Icon type="loading" style={{ fontSize: 40 }} spin />} /></div>}
-      <Route path="/programs/:projectId" exact render={() => {
-        if(!loading && results.length > 0) { return [
-          <Select allowClear optionFilterProp="children" value={countryFilter} onChange={handleCountryFilter} mode="multiple" placeholder="All countries" className="country-filter" dropdownMatchSelectWidth={false}>
-            {countryOpts.map(opt => <Option value={opt}>{countriesDict[opt]}</Option>)}
-          </Select>,
-          <Collapse defaultActiveKey="0" destroyInactivePanel onChange={handleResultChange} accordion bordered={false} expandIcon={({ isActive }) => <ExpandIcon isActive={isActive} />}>
-            {results.filter(filterCountry(countryFilter)).map((result, index) =>
-              <Panel
-                key={index}
-                header={(
-                  <StickyClass offset={20}>
-                    <h1>{result.title}</h1>
-                    <div><i>{result.type}</i><span>{t('nindicators', { count: result.indicatorCount })}</span></div>
-                  </StickyClass>
-                )}
-              >
-                <Result programId={params.projectId} id={result.id} {...{ countryFilter, results, setResults }} />
-              </Panel>
-            )}
-          </Collapse>
-          ]
-        }
-        if (!loading) return <Redirect to={`/programs/${params.projectId}/editor`} />
-        return null
-      }} />
-      <Route path="/programs/:programId/hierarchy/:projectId?" render={(_props) =>
-        <Hierarchy {..._props} canEdit={canEdit} program />
-      } />
-      <Route path="/programs/:projectId/reports" render={() =>
-        <Reports programId={params.projectId} />
-      } />
-      <Route path="/programs/:id/editor" render={({ match: {params}}) =>
-        <Editor {...{ params }} program />
-      } />
-      <Route path="/programs/:id/approvals" render={({ match: {params}}) => <Approvals {...{params}} />} />
-      {/* <div id="chartjs-tooltip" /> */}
-      <div id="bar-tooltip" />
-      <div id="disagg-bar-tooltip" />
-    </div>
-  )
-}
-
-export default connect(
-  ({ editorRdr: {section1: {fields: {title}}}, userRdr }) => ({ title, userRdr }), actions
-)(Program)
